Throw when our delay wrapper is not wrapping a Delay
When the wrapped interceptor of
DelayPropagatingContinuationInterceptorWrapper does not implement Delay,
the implementation did an unchecked cast from Dispatchers.Default to
Delay. But Dispatchers.Default doesn't implement Delay so that cast
throws.

Instead, throw a more descriptive error and require that the wrapped
interceptor implements Delay.

This also means we can make the argument non-nullable.

Bug: N/A
Test: presubmit
Relnote: N/A

Change-Id: I0258d1c27cf1b87bc1ca27f5e17ecfa732c6ab13