Add SharedUiContainer that extends ViewGroup
A new container that will be used to maintain a session connection with
sandboxed SDKs is introduced. This change focuses on the layout logic
of the container, i.e. onMeasure() and onLayout() implementation to
support container's children layout.

All views added to the container are placed in the top left corner plus
padding, with newer children overlaying older ones. Padding is also taken into
account when measuring child views. Container's size
is determined by the parent's MeasureSpec, padding settings and its
largest child. Children that are View.GONE are not measured and/or laid
out.

Session management, asset registration logic and documentation will be
provided in the next change.

Relnote: "Added SharedUiContainer that can host client-owned and provider-owned UI
and lays out its children in top left corner"
Test: build, SharedUiContainerTest
Bug: 263460954

Change-Id: Ia7310cc5be45e54912aebee0669aab6152de59bd