Add Autofill highlight capability
This CL builds off of aosp/3229628 and introduces a `compositionLocal` by which developers may modify the hue of the autofill highlight that indicates when filling has successfully completed. See go/compose-autofill-docs for more details.

Bug: 349386998
Test: performAutofill_customHighlight_BTF() and others
RelNote: Introduce `CompositionLocal` that can be used to modify the hue of Autofill's successful filling highlight.
Change-Id: I320926246d1811c2c974a6793f48941907c33afe