Ensure PredictiveBackHandler respects late enabled changes
It seems that it is possible to get into a case where after a Predictive
Back gesture has been started, the handler is disabled and instead of
cancelling the gesture, the PredictiveBackHandler is executed as if it
were still enabled.

We can change PredictiveBackHandler to take into account any late
changes to the enabled state and be sure to keep any pending callbacks
from executing at all.

This means if you are on the last destination in the
back stack and you some how get in this state, the gesture arrow will
show, but will not work.

Test: Added PredictiveBackHandlerTest
Bug: 340202286
RelNote: "NavHost will no longer call an IndexOutOfBoundsException when
the PredictiveBackGesture is enabled."

Change-Id: Ie3301e4a06593101bc48c447f3a9505a1de5d6a3