Add media3effect to test app with device tests
Add the new Media3Effect to the view test app. The new fragment contains a sliding bar that adjusts the brightness of the preview.

Also add device tests that verifies that when the sliding bar is at max value, the output is white; when it's at min value, the output is black. That way we can test the media3 effect is working correctly reguardless of the camera input, which is hard to guarantee without e2d testing gears.

Bug: 364293282
Test: manual test and ./gradlew bOS
Change-Id: I8fd83a5af19d4fedbe84e2e7d779a9634c52388c