Share TestScheduler in tests that use TestScope
When testing using the TestScope and its TestScheduler it is important to share the same TestScheduler in the test function and in the objects used in the unit test. This is done by using runTest() from the test class TestScope.

See https://developer.android.com/kotlin/coroutines/test#creating-your-own-testscope

This CL also fixed an issue in RoomDatabase explicit close() call. It removes the compatibility mode check to no-op if the SupportSQLiteDatabase was never opened. However this check prevented the Coroutine Scope and other cleanups from happening and is not needed since a close() call on a SupportSQLiteOpenHelper or SupportSQLiteDatabase that are not open are no-op.

Test: Existing
Change-Id: Iecb0af28dd847ee0d68e83eed98a2e49359290da