Handle sandbox death inside all SandboxedSdkViews
Currently the error is not handled gracefully in all
the SandboxeSdkViews inside the UiLib test app leading
to a red SandboxedSdkView background instead of the
Remote Process Died error message in some of the views.
This change will make sure that the error is handled
gracefully for all the SandboxeSdkViews with the error
message. Also, makes the error message bold Italic and
Red in color to be clearer.

Bug: 362489332
Test: UiLib Test App
Change-Id: I10b3a606ff044f2a27ddc883b1e68d1c7c982e11