Add Autofill navigation API and samples
Expands on the semantic autofill manager to add the `commit` and `cancel` APIs that will allow developers to notify the autofill framework when an Autofill session is completed or canceled. This is part of the saving user journey and will allow users to save new login/credential information. See go/compose-autofill-docs for more details on how to use the APIs.

Test: autofillManager_notifyCommit(), autofillManager_notifyCancel()
Bug: 333102566
Bug: 349386998
RelNote: Expands Autofill manager to include `commit()` and `cancel()` APIs that help users save newly entered credentials.
Change-Id: I2da00b7050b12af7b978f2f52fe27dc74a6fd8f2