Add lint warning for data class companion to popBackStack
If you have a data class that is marked as @Serializable you should not
attempt to pass the object to popBackStack. You can pass an instance of
the object if you have the proper arguments available, but in the case
that you don't you should be using the version of the popBackStack that
takes a reified class instead.

This change adds a lint rule to catch cases of passing in a Companion
object and also offers a quick fix.

RelNote: "Navigation now offers a new lint rule to warn against passing
@Serialized data class objects to popBackStack. It will suggest using
the reified class version instead."
Test: Added tests along with rule
Bug: 358095343

Change-Id: Ief1614a3ff15937b173536a43ba3f0de3c4d8b8a