Cache WindowAreaController when created
Removes the need to handle the scenario of
trying to perform an operation on a
WindowAreaInfo with a new WindowAreaController.

This will create a new WindowAreaController when
it is being created for the first time, thenreturn
the cached WindowAreaController anytime getOrCreate
is called after that.

Bug: 390230656
Test: WindowAreaControllerImplTest
Change-Id: Ibae21fb69c8296e84d8bfb11cbf0a346e295a695
Merged-In: Ibae21fb69c8296e84d8bfb11cbf0a346e295a695