Avoid redundant nullability annotations in GenCode.
This issue was first encountered in a google3 test breakage. This was due to the room-paging module that was recently migrated to be KMP, the process in which we had added a new API convertRows(statement: SQLiteStatement, itemCount: Int) as a KMP replacement to the previous convertRows(cursor: Cursor). The Int parameter `itemCount` in the Java generated code is getting annotated @NonNull, which is incorrect as a primitive cannot be nullable.

The incorrect generated code was due to the `addParameter` API in JavaFunSpec annotating all generated code parameters with NULLABLE or NONNULL without checking for primitive types. This check has now been added, and a Room generated code test that verifies the issue has been added for the room-paging case.

Bug: 362512509
Test: DaoWriterTest.kt
Change-Id: I76e0d109c91376fae12532911cf359dcd0ebeced