Refactor navigating with string route
Refactor navigate() overload for string routes to use the new matchRoute APIs.

Navigating with safe args object is refactored to use navigate() with string route. This is because navigation with objects converts the object to a route before navigating.

Existing tests have been refactored - tests that navigate with string route but converts the route to the URI will no longer do the conversion. This is because navigating to route with URI is no longer possible.

Test: existing tests
Bug: 360084191
Change-Id: I9ac968603130274e0624363322343dfb274400cf