Fix thread usage in RoomTrackingLiveData
In aosp/3205550 Room's LiveData implementation was converted to Coroutines and in aosp/3238929 the main thread usage was fixed. But a difference from the original executor implementation was that refresh() is scheduled either on the query executor or transaction executor it should not thread-hop in between like it was changed in aosp/3238929 using withContext. This fixes the threading but launching in refresh() coroutine in the right thread pool.

Bug: 361216129
Test: Manually validated test in b/361216129 with snapshot version.
Change-Id: I3b94bfdd3c324c4728542e65b572b526681dd5f8