Don't emit null check for RxJava query types.
When Kotlin codegen is turned on, type argument nullability is important and Room emits null checks for when the result is not null and the query result is empty. However, for RxJava types specifically the type adapter should not emit null checks since the type arg should always be non-null and when the query result is empty, no emission is done as RxJava does not allow null values

This fixes an issue where users with non-null type args of Flowable and Observable that switched to Kotlin codegen where forced to update the nullability (unnecessarily) or else risk a query throwing an exception.

Bug: 320623557
Test: RxJava2QueryTest
Change-Id: I7e46d01a29453a14460d212025eac818ff1b6d68