Affect bounds when insetting dialog content
Fixes: 363916110

Reviewer found that the Bounds were not being affected by the
animation handler in a different part of the code. This also affects
Dialogs. I fixed it here so the bounds can be updated when the
content is inset.

This also fixes a problem when the dialog is offset from the
screen.

Test: new test
Change-Id: I8b5a68ac6d7d63903503f989639c01bc67371dd6