Updating documentation for Material Dialog
Dialog does not call onDismissRequest in response to the showDialog parameter being set to false. We have updated the documentation to reflect that. On the other hand, the outro animation is triggered when setting showDialog to false, which is correct behavior.

Moving the existing LaunchedEffect so that onDismissRequest would have been called when showDialog was changed to false (as suggested in the bug report) had unintended consequences - in some cases, onDismissRequest was called multiple times. That would have been a breaking change for existing usage.

Bug: 357829790,
Test: DialogTest
Relnote: "Updating Material Dialog description that onDismissRequest is not be called after show flag is set to false"
Change-Id: Ifd8d6d17abe8b5316a37b56cf08df4226426a058