Enable Parameterized Tests for Camera-camera2 and Camera-pipe on Presubmit
This change enables the following parameterized tests
to run on both camera-camer2 and camera-pipe integrations
during presubmit.

This change will increase the total execution time
of presubmit tests.
There might be an initial increase in the failure rate
as we identify and fix any compatibility issues between
the tests and the two integrations.

Test: BasicUITest, Camera2InteropIntegrationTest, CameraDisconnectTest, CameraInfoDeviceTest, CameraXInitTest, CameraXServiceTest, CaptureOptionSubmissionTest, ExistingActivityLifecycleTest, FlashTest, FocusMeteringDeviceTest
Bug: 361248535
Change-Id: I90a537439b2c423b62768059e6c5972109322c2f