Use compileOnly for AGP and remove KGP dep in SafeArgs Gradle Plugin
Using compileOnly for other plugin dependencies prevents consumers of the SafeArgs Gradle Plugin to have their plugin upgraded or conflict with buildSrc. This does require that the plugin not use any class from those other plugins unless they are applied (by id) or else risk a ClassNotFoundException. In the case of KGP since we only need to check if the project is using Kotlin, we can do that with just the id and without referencing any KGP classes, hence we can remove the dependency completely.

Also updated the AGP under test to 8.1.0 to avoid conflict with too new KGP and added required Android namespace to test projects.

Bug: 365378970
Test: ./gradlew :navigation:navigation-safe-args-gradle-plugin:test
Change-Id: I7628797a872d275242df3d64df09005b02ff621c