Refactor navigation common lint tests
Move stubs from navigation-runtime-lint to lint-common.

Refactor navigation-common-lint tests to use the common code.

Removed the test case for only source code. We should always be using
the bytecode for our library stubs. Our test code stubs source is only
beneficial if the source is out of sync with the bytecode, and the
helper functions notifies of any changes to the source code that the
bytecode should be updated.

Since we are now only testing with bytecode, we can remove the
paramaterized tests and simplify to regular junit tests.

Test: all existsing tests have been updated and pass
Change-Id: I2e0f1fa719bb285f857e3467499d40566845e550