Revert^2 "Adopt common Drag&Drop API for multiplatform"
This reverts commit 95640577b703090cd2ffb484e558deb6e09372eb.

Reason for revert: Temporary restored old API in foundation

Relnote: `Modifier.dragAndDropSource` was adopted to multiplatform. Drag gesture start detection is now performed by Compose itself. `Modifier.dragAndDropSource` does not allow customizing the start of the gesture anymore and instead exposes a lambda that is invoked with an offset when a drag session starts. The old overload has been deprecated. To use the new overload, specify a name for the lambda parameter or explicitly specify the offset parameter in the lambda. See `DragAndDropSamples.kt`.
Relnote: `DragAndDropModifierNode` is deprecated in favor to `DragAndDropSourceModifierNode` and `DragAndDropTargetModifierNode`
Test: existing tests + manually
Change-Id: I9f54594cd8b8441321f45034fd248aaa85b9f2c8