Remove confirmValueChange from AnchoredDraggable
confirmValueChange's contract is hard to reason about and conceptually, confirmValueChange does not encourage good user experience.
Additionally, developers used confirmValueChange as a callback, even though it does not provide any invocation guarantees.

For example, a bottom sheet that is dragged 90% from Expanded to Hidden and the flung will call confirmValueChange. confirmValueChange returning false will make the sheet move back to the Expanded anchor. Instead, the Hidden anchor should be excluded from the set of anchors if it is not allowed to be moved to. This means an OverscrollEffect supplied to Modifier.anchoredDraggable will be applied, indicating to the user that their intended action is disallowed.

Relnote: Deprecated AnchoredDraggableState's confirmValueChange. Instead of vetoing state changes, disallowed anchors should not be in the active anchor set, and an OverscrollEffect should be used to indicate the unavailability of the requested action.
Test: Keeping existing tests until we remove the deprecated overloads.
Change-Id: Ia717fb7dba3db65609a933a55b164d2186e25806