AutoSize.StepBased: Fix getFontSize() bug
In certain cases the getFontSize() was returning a font size that was stepSize greater than the correct font size.

This has been fixed by rounding down min instead of current as we know this doesn't overflow (or if it does, it is equal to minFontSize).

This does not impact the rest of the logic because we round down min so that it's divisible by stepSize when subtracted by minFontSize, then add stepSize, which gives a font size between min and max, and we check for overflow here. If we add stepSize again, we are guaranteed to be above max as the difference between max and min is strictly smaller than stepSize, thus guaranteeing us that we check every potential font size we have left.

Test: Added new test case to AutoSizeTest that failed before and passes now.
Change-Id: Id6cd8ef264264b1233a7df6ff9856ecc2b111745