Consider seeking when cancelling transition effects
Currently if you run a seekable transitions, when the operation is
cancelled, we always attempt to reverse the transition. The problem is
that there are cases when we actually want to finish the transition.

If the operation is cancelled by another incoming operation, (i.e. there
was a fragment transaction) we have been interrupted and should try to
reverse the transition. However, if the operation is cancelled by
forceComplete (i.e. fragment manager state changes, another back event,
executePendingActions, etc), we want to finish the transition.

This change looks at the seeking status of the operation, inside the
effect so that effects can now cancel with and without seeking. This
allows the TransitionEffect to consider seeking when deciding whether to
reverse the transition.

RelNote: "Fixed an issue where quickly pressing system back or doing
fast gesture back will cause Fragments to crash."
Test: Added test in transition
Bug: 364804225
Fixes: 364914301

Change-Id: Ibc038d8db7c3e7903a4dc8bfa556883705d27ee8