Supporting ByteBuffer in non-Android & non-JVM platforms.
Since ByteBuffer is a JVM-only API, it was erroneously used in generated code in non-JVM platforms. It has been replaced altogether with a new ByteArrayWrapper class which holds a byte array and replaces the use of a ByteBuffer by implementing equals() and hashcode().

Bug: 367205685
Test: BaseQueryTest
Relnote: Create internal ByteArrayWrapper class to support Relations with ByteBuffer in non-Android & non-JVM platforms.
Change-Id: I755433b8d624c88a6d645d40cf845d9bd3123398