Ensure on screen HardwareBuffers get released.
If a BufferPool is closed, it only releases HardwareBuffers
that are not currently being used. As a result it is the
responsibility of the consumer of the pool to release the buffer.
Ensure that currently presented buffers are also released in the
event the dependencies are being released due to a resize of
the underlying parent SurfaceView.

Fixes: 367539482
Test: manual tracing of memory profiler
Change-Id: I88aec23c223b20b2ce0ee381d71296561150e9c4