Correct order of overloaded constructors
Constructor overloads generated by Metalava to work around Psi issue
https://youtrack.jetbrains.com/issue/KT-57537 were generated in the
wrong order because they were generated while writing the signature
which meant that they were not sorted like all the constructors in the
model. That has been fixed in the latest version of Metalava and this
change updates the signature files to match.

Relnote: N/A
Bug: 367623741
Test: ./gradlew
Change-Id: I01212318aa18152bca43a283add6987357c1fad7