Rename `invisibleToUser()` to `hideFromAccessibility`.
After some investigation in go/a11y-experimental-invisible, it seemed that `invisibleToUser()` had an ambiguous name and its main function was to mark nodes as unimportant for accessibility services. This CL renames the API as per go/stabilize-invisibletouser. Note that this property is either present or not; it isn't a boolean that can be set to `true` to mark something as important for accessibility.

This CL deprecates `invisibleToUser()`. After removing usages of the old API in g3, then we can proceed with removing it in our library.

Note that this was added and then reverted; this newest version provides backwards compatibility for the deprecated API.

Fixes: 347749977
Fixes: 367714648

Test: modified existing tests and added `populateViewStructure_hideFromAccessibility()`
RelNote: Rename `invisibleToUser()` to `hideFromAccessibility`. Its function remains the same. See documentation for more details.
Change-Id: Ib43a30cc748b503b4e70551cdfe55c30118870b1