DatePicker samples update
Update the DatePicker samples to include inner scrolling when the
content does not fit the height (i.e. in landscape mode on some
devices).
The current implementation does not have a dedicated lanscape layout,
and we still recommend using a DisplayMode.Input at the state whenever
there is not enough room, however, this approach can also be applied to
handle these cases.

Bug: 335223055
Test: Manual test at the Catalog app
Change-Id: Ifce03fab1f8ac03f78b60aa6415a473d31299216