Remove special-casing for decorFitsSystemWindows for API 35+
Fixes: 364492593

Previously, when the app targeted API 35 and above, dialogs would
automatically have decorFitsSystemWindows = false. This caused some
dialogs to draw behind the navigation bar unnecessarily and would
change the navigation bar color. It also has strange effects when
the dialog does not use the entire screen size and the device
is in landscape mode.

This CL removes that special handling and decorFitsSystemWindows
acts the same on all API versions now.

Test: ran broken use case in the broken app
Test: new test for height irrespective of target API

Change-Id: Ibc94bd0bdfe6645e2e8b92dfbe39c93fece38a2f