Use smaller size for Dialog windows
Fixes: 364492593

When dialog contents are smaller than the screen size, the
dialog's window should not consume the entire screen. When
it does, the navigation background's color may not match
the expected style.

This CL uses WRAP_CONTENT for height, but when
decorFitsSystemWindows and usePlatformDefaultWidth
are false and the dialog needs more space,
MATCH_PARENT is used. The content is centered within
the Window in all cases.

Test: manual testing, testing on broken app
Test: new tests
Change-Id: Iffb87ff0b4637361b23b4867e903a59949bd8180