Enable compiling jvmStubs target tests
In aosp/3211838, :compose:runtime:runtime:compileTestKotlinJvmStubs task did not run in presubmit even though jvmStubs.runTests was set to true. This lead to androidx target failing in postsubmit.

This was becasue we did not add compilation of jvm stuvs to the allHostTests anchorTask. Now, we compile the test jvm stubs, but only run when the flag is set to true

TESTSED: Ran ./gradlew :compose:runtime:runtime:allHostTets and the comppileTestKotlinJvmStubs and jvmSubsTest tasks ran. Then disabled running the tests and re-ran gradle command, and only the compile test ran (as expected)

BUG: 362976596
Change-Id: I911ca902e1a41e77941878efaae9dc1067e3e3ba