Require API 34 device for accessibility checks
Accessibility checks for Compose only work on API 34, and it's currently
not implemented in Robolectric at all.

Add @RequiresApi(34) to all related API, and log a warning when enabling
accessibility checks on Robolectric. A warning is logged rather than
throwing an exception, because a future version of Robolectric will
likely add support. This enables users to only upgrade Robolectric,
without a hard requirement to upgrade `ui-test`.

Fixes: 362771513
Test: presubmit
Relnote: N/A

Change-Id: I50ca36bb6109b0be17bdf1fab50c82a9bd3304ec