Remove ReusableContentHost from lazy items
ReusableContentHost was introduced to ensure that reused items keep their content. This deactivation, however, is not synchronised with SubcomposeLayout wrapper, making it possible to access and measure deactivated content. This change will cause deactivated content to be sometimes lost, leading to a slightly worse performance when lazy items are changed.

The test only reproes with LookaheadScope, but it is possible that in certain conditions the LazyColumn can trigger a similar condition on its own.

Fixes: 368429360
Change-Id: Id6e6087b1ac571d991de535d85c465936a3e4c1c