Change mechanism to prevent backwards prediction
Under specific circumstances it was possible to have very large
predictions; the root cause is an unfortunate interaction between
the prediction ratchet not capping prediction in the future, and
the prediction event time being computed incorrectly.

This change will correct both of these issues: the prediction will
now not go beyond the previous one unless specifically requested,
and the event time calculation is now more accurate.

Bug: 369330439
Test: gradlew :input:input-motionprediction:test
Change-Id: Ifbf492418e3a15ff4743526a394c2da7c3941268