Ignore some tests on API 22 and below
Somehow running these two tests that just create a StaticLayout underneath fails some random tests (almost if there's some global state). When running those affected tests in isolation, they pass.

These tests that are ignored are actually testing some logic inside Compose and we don't branch for API there, so it should be pretty safe to not runt he tests on API 21 and 22. Also note that in general ellipsis support for start and middle position on those APIs is somewhat incorrect on the platform anyway.

Fixes: 364169257, 366492841, 366492853, 366493166
Test: compose:ui:ui-text:cC on API 21, 22, 35
Change-Id: Ia046674e96d17e2fdfb7ef34c85fbe19af6350ba