Make background color of SandboxedSdkView White
In UiLib test app, the background color of the
SandboxedSdkView is Red by default. We should
convert it to color white so that it can be
distinguished from the red viewability rectangle.
This will be especially useful when the padding
is applied to the SandboxedSdkView to distinguish
the rectangle and the background.

Bug: 369548816
Test: UiLib Test App
Change-Id: I812a4296ce2f5d6c01b835661bfa3971664db77a