Add Material3.EdgeButton component
Since this component's inner content can be icon or text and we want to
have opinionated defaults for them, split it to 2 functions:
- iconEdgeButton
- textEdgeButton

with additional "container" that accepts style for those two.

It is slightly different from Compose who only has EdgeButton, but it's
align with "round" Button.

Bug: 368272637
Test: As samples, unit will be after testing library
Relnote: "Added Material 3 edgeButton component."
Change-Id: I82b94222093e642c6961fa70139d21bcbf40dd2a