Deprecate GlobalAssertions
The experimental API to add GlobalAssertions did not meet our standards
for API design and we couldn't find justification to rework this API to
make it meet our standards.

Please file a feature request with your use case if you want this or a
similar API in the future at https://issuetracker.google.com/ in the
component "Jetpack Compose -> UI Libraries -> Testing".

Fix: 296400284
Test: existing tests
Relnote: "Deprecated the experimental GlobalAssertions API. Its intended
use was to run accessibility checks, see
[`enableAccessibilityChecks()`](https://developer.android.com/reference/kotlin/androidx/compose/ui/test/ComposeUiTest?hl=en#enableAccessibilityChecks())
for that purpose instead."

Change-Id: I50aa509889849f8772724f46d5849920cf4fd660