Date pickers Locale support
This change improves how the DatePicker and DateRangePicker handle
locales when a Locale that was set by the date picker's state wasn't
consistently applied to all its parts, leading to potential
inconsistencies in things like date formatting, weekday names, the
first day of the week, etc.

The DatePickerState and the DateRangePickerState were modified to
include and expose a Locale value that can be used by the various parts
of the UI.

There are two ways to create the state:
- Using remember functions: This will default to your device's locale
  and is the recommended way.
- Using DatePickerState or DateRangePickerState directly: This allows
  you to specify a custom Locale that will be applied most of the date
  picker parts, but not all. Note that the title and the headline's
  default text will need to be adjusted to support a specific Locale in
  case a specific one is set.

Fixes: 326490763
Fixes: 321657276
Test: additional date picker and input tests
Relnote: "Updates to the date pickers to ensure consistent Locale
usage throughout the DatePicker and DateRangePicker when setting a
Locale directly through a DatePickerState or a DateRangePickerState.
Note that when setting Locales directly to the state, it's up to you to
ensure that the title and headline texts are localized accordingly, as
their default text will still be applied according to the default
platform Locale."

Change-Id: I370735477b6adae8628fb967b6aa6c6138216883