Update primitive collection check to work with AGP update
The upcoming AGP update pulls in a K2 change which removes constructors from the psi representation of value classes (see b/369846185 for more details).

The PrimitiveInCollectionDetector check relied on value class constructors to determine whether an inline value class was really a primitive. This looks at fields instead to find the type of the value class.

Bug: 369422908
Test: PrimitiveInCollectionDetectorTest, `./gradlew wear:compose:compose-foundation:updateLintBaseline` runs with no diffs without the AGP update and with the AGP update (without this change and with the AGP update, the baseline would be updated to reference object collections instead of int collections).

Change-Id: If3c4c1f0a4d74c60e897f82a84c997827a69a42c