Move to jarjaring health:health-services-client proto dependency
We split the protos away from the client library as done elsewhere in AndroidX to repackage the protobuf runtime namespace to a custom one as the runtime lib is not compatible between versions.

Also, the RepackagePlugin needs the JAR task to exist, which only exists in the JavaLibraryPlugin and KotlinBasePlugin. Both these arent applied in the health-services-client project.

TESTED: Ran ./gradlew :health:health-services-client:publish and it succeeded
Checked .module file generated and it contained a dependency on the proto

BUG: 370114848
Change-Id: If664a23e25f717e72a11ce6751a6b6f9bf74a07a