Only call platform MR#selectRoute upon app request
This change prevents calls to platform MR#selectRoute as a result of MR
callbacks, as those are causing spurious route selections. This can
happen, for example, when:
1. A message to select the phone speakers is posted to a handler as a
   result of a MR callback.
2. A BT devices is connected.
3. The message from step 1 executes. At this point the framework sees
   the selection of built-in speakers while a BT device is available,
   causing the playback on BT to be overriden by the built-in speakers.

In the longer term, we want to stop depending on
android.media.MediaRouter for Bluetooth route management, in favour of
using MediaRouter2 instead. But given this can cause API capabilities
regressions for apps without some permissions, we need more time before
commiting to those changes. In the meantime, this change should be a
safe way to mitigate incorrect built-in speaker selection bugs.

Bug: b/294968421
Test: ./gradlew :mediarouter:mediarouter:connectedAndroidTest
Test: Manually using the system routing activity in the sample app.
Change-Id: Ib9bf0a0f96ce359ac2cbc3e22ebf85df563ae5b6