Add FeatureDetector
Adds a utility file that converts a list of cubics to features. Straight cubics are converted to edges, others to corners. Subsequent cubics may be merged to one feature if their points align / are extremely close. To reflect the different characteristics cubics can have (be straight, be smooth from their predecessor, etc) a few extension methods on Cubic have been added. Its mostly the Cubics methods that are being tested for now, as the detectFeatures method will play a significant part in the public API testing.
Bug: 372266653

Relnote: Add FeatureDetector
Test: Unit tests for cubic characteristics
Change-Id: If68edad7e838351be9e8e8be9e5977adb4f4f7e3