Date picker input mode validation error height
Fix the bottom padding at the DatePicker in an input-mode when a
validation error displays an error text.
The fix ensures that the height of the component stays unchanged when
that happens.

Fixes: 280462363
Test: Added to DateInputTest
Relnote: "Fixed an issue with the DatePicker in input mode where
validation errors could cause the component's height to change"

Change-Id: I2e229559964acd9522e8a696bb0e89698c20bdd9