Update how LICENSE files are checked in prebuilts/androidx/external
Move to a task that walks the file tree of prebuilts/androidx/external
to validate whether all of the libraries with a .pom file have a LICENSE
or NOTICE present along side it.

As-is, we have some missing entries, so adding a baseline that we can
reduce as a follow up.

Test: ./gradlew :checkExternalLicenses
Change-Id: I848d4efdc439e3016466ac290bf94d27dacc8825