AlertDialog: addressing API Review Feedback
Some api changes were suggested by the API council. The main change was an addition of `AlertDialogContent` and `Dialog` apis for easier customization . While `AlertDialog` remains a preferred usage for notifying a user, it is now possible to use`AlertDialogContent` with `Dialog` if custom animation or other changes are required.

Other changes include:
- Enhanced documentation for the `title` parameter to provide clearer guidance.
- The addition of `modifier` and `color` parameters within AlertDialogDefaults members for increased customization
- The inclusion of `contentDescription` for default icons to improve accessibility.

Bug: 371046842
Test: AlertDialogTest.kt
Relnote: "AlertDialog api was updated with additional `AlertDialogContent` and `Dialog` composables. `Modifier` and `color` parameters were added to `AlertDialogDefaults` members as well."
Change-Id: I4eb71abb2f29170154475aa707c915b55b2fec0e