Avoid loading androidx.graphics.path for non-Android platforms
The androidx.graphics.path library is build on the NDK and only works in
Android. Avoid loading it on host JVM runtimes such as LayoutLib and
Robolectric. This will make LayoutLib and Robolectric support
PathIteratorImpl on SDK 34 and above.

Note that the PathIteratorPreApi34Impl native methods will be unbound,
and trying to invoke them will result in an UnsatisfiedLinkError.
However, on API 34+, PathIteratorApi34Impl is used and defers to the
framework implementation, so this is an improvement over the prior
state.

Test: graphics path tests
Bug: 354621198
Change-Id: I32b2a6aa0d8e029db36d3d8782e789c330ba56dd