Add semantic CurvedModifier
Added a first version of semantics to CurvedLayout.
This includes a CurvedModifier to specify content description
and/or traversal index, that can be set on any curved text or
curvedComposable (that are the only curved elements emiting
composables).
As the CurvedLayout itself is a Composable, semantics can be set
on it directly using Modifier.semantics.
Relnote: "Added CurvedModifier.semantics, currently supporting
content description and traversal index"
Test: N/A
Bug: 373334637

Change-Id: I0b0931dfb487af484758959dd0f036bc10fd58ba